Split validate command
