+++ cmd/util-db/validate/validate.go
@@ -1,20 +1,4 @@
-// Copyright 2024 Fantom Foundation
-// This file is part of Aida Testing Infrastructure for Sonic
-//
-// Aida is free software: you can redistribute it and/or modify
-// it under the terms of the GNU Lesser General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
-//
-// Aida is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-// GNU Lesser General Public License for more details.
-//
-// You should have received a copy of the GNU Lesser General Public License
-// along with Aida. If not, see <http://www.gnu.org/licenses/>.
-
-package db
+package validate
 
 import (
 	"bytes"
@@ -27,8 +11,8 @@ import (
 	"github.com/urfave/cli/v2"
 )
 
-var ValidateCommand = cli.Command{
-	Action: validateCmd,
+var Command = cli.Command{
+	Action: validateAction,
 	Name:   "validate",
 	Usage:  "Validates AidaDb using md5 DbHash.",
 	Flags: []cli.Flag{
@@ -36,8 +20,8 @@ var ValidateCommand = cli.Command{
 	},
 }
 
-// validateCmd calculates the dbHash for given AidaDb and compares it to expected hash either found in metadata or online
-func validateCmd(ctx *cli.Context) error {
+// validateAction calculates the dbHash for given AidaDb and compares it to expected hash either found in metadata or online
+func validateAction(ctx *cli.Context) error {
 	log := logger.NewLogger("INFO", "ValidateCMD")
 
 	cfg, err := utils.NewConfig(ctx, utils.NoArgs)
